m10_2_tool_calling/requirements.txt
# Module 10.2: Tool Calling & Function Execution
# Python dependencies

# Core Framework
pydantic==2.6.0
python-dotenv==1.0.1

# Sandboxing & Safety
RestrictedPython==7.1

# Retry & Resilience
tenacity==8.2.3

# Database
psycopg2-binary==2.9.9

# Slack Integration
slack-sdk==3.26.2

# LLM & Embeddings
openai==1.12.0

# HTTP & API
requests==2.31.0

# FastAPI (for app.py)
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5

# Monitoring (optional)
prometheus-client==0.20.0

# Utilities
numpy==1.26.4
